requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4
aioitertools==0.13.0
aiosqlite==0.22.1
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.12.1
appnope==0.1.4
arize-phoenix==12.33.1
arize-phoenix-client==1.28.0
arize-phoenix-evals==2.8.0
arize-phoenix-otel==0.14.0
asttokens==3.0.1
async-timeout==3.0.1
attrs==25.4.0
Authlib==1.6.6
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
Bottleneck==1.6.0
build==1.4.0
cachetools==7.0.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
chardet==3.0.4
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
cmd2==1.5.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
commonmark==0.9.1
cross-web==0.4.1
cryptography==46.0.4
debugpy==1.8.20
decorator==5.2.1
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durationpy==0.10
einops==0.8.2
email-validator==2.3.0
executing==2.2.1
fastapi==0.128.0
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fsspec==2026.1.0
ftfy==6.3.1
geographiclib==2.1
geopy==2.4.1
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-aiplatform==1.135.0
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.8.0
google-cloud-vectorsearch==0.3.0
google-crc32c==1.8.0
google-genai==1.61.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphql-core==3.2.7
greenlet==3.3.1
groq==0.37.1
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.16
idna==2.10
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.0.0
jiter==0.12.0
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kubernetes==35.0.0
langchain==1.2.7
langchain-core==1.2.8
langchain-google-genai==4.2.0
langchain-google-vertexai==3.2.2
langchain-groq==1.1.2
langchain-ollama==1.0.1
langchain-tests==1.1.3
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.7
ldap3==2.9.1
lia-web==0.3.1
Mako==1.3.10
markdown-it-py==4.0.0
markdown2==2.5.4
markdownify==1.2.2
marker-pdf==1.10.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.10.0
numexpr==2.14.1
numpy==2.4.2
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==1.109.1
opencv-python-headless==4.11.0.86
openinference-instrumentation==0.1.44
openinference-instrumentation-langchain==0.1.58
openinference-semantic-conventions==0.1.26
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.6
ormsgpack==1.12.2
overrides==7.7.0
packaging==25.0
pandas==3.0.0
parso==0.8.5
pdftext==0.6.3
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.5.1
pluggy==1.6.0
ply==3.11
portalocker==3.2.0
posthog==5.4.0
pre_commit==4.5.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.26.7
pymupdf4llm==0.2.9
pypdfium2==4.30.0
pyperclip==1.11.0
PyPika==0.50.0
pyproject_hooks==1.2.0
pystache==0.6.8
pytest==8.4.2
pytest-asyncio==1.3.0
pytest-benchmark==5.2.3
pytest-codspeed==4.2.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.16.2
RapidFuzz==3.14.3
referencing==0.37.0
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.17.17
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlean.py==3.50.4.5
stack-data==0.6.3
starlette==0.50.0
strawberry-graphql==0.287.3
surya-ocr==0.17.1
sympy==1.14.0
syrupy==4.9.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.2
traitlets==5.14.3
transformers==4.57.6
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
validators==0.35.0
vcrpy==8.1.1
virtualenv==20.36.1
watchfiles==1.1.1
wcwidth==0.5.3
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
